progress update at Algorithmic Heights
Focusing more on algorithms in general than on specific applications in Bioinformatics. It's a fun challenge for now while I'm using only the things I understood from the task text and anything that I can remember in relation to the task which could help me solve it.
Might have to consult literature or sample code later but I'll try my best not to, first focusing on a working solution rather than immediately providing the best one that can be found online.